Resolved the final tweaks.
